+++ changelog.md
@@ -1,5 +1,6 @@
 ##### Version 0.16-106
 - Initial port to 1.16
+- Added mineral vein, crusher & arc furnace integration for the new Nether materials
 - Also includes all changes from 0,15-105
 
 ##### Version 0.15-105